Refactor [#9] Timer 관련 패키지 이동
application으로 Service 이동
