+++ morib/src/main/java/org/morib/server/domain/timer/application/ClassifyTimerServiceImpl.java
@@ -1,15 +1,11 @@
-package org.morib.server.api.homeViewApi.service.classify.timer;
+package org.morib.server.domain.timer.application;
 
 import lombok.RequiredArgsConstructor;
-import org.morib.server.domain.category.infra.Category;
 import org.morib.server.domain.task.infra.Task;
 import org.morib.server.domain.timer.TimerManager;
-import org.morib.server.domain.timer.infra.Timer;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
-import java.util.List;
-import java.util.Map;
 
 @RequiredArgsConstructor
 @Service
@@ -17,7 +13,7 @@ public class ClassifyTimerServiceImpl implements ClassifyTimerService{
     private final TimerManager timerManager;
 
     @Override
-    public Map<Task, Timer> classifyByTask(List<Task> tasks) {
-        timerManager.classifyByTask();
+    public int classifyTaskByDate(LocalDate date, Task task) {
+        return 0;
     }
 }
